// web/src/__generated__/schemas/CustomerActorCustomerView.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { RealmView } from './RealmView';

export type CustomerActorCustomerView = { 
  customerId: string; 
  name: string; 
  alias?: string | null; 
  realms?: RealmView[]; 
};