// web/src/__generated__/schemas/CustomerActorNewCustomerRequest.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

export type CustomerActorNewCustomerRequest = { 
  email: string; 
  password: string; 
  realmName: string; 
  alias?: string | null; 
  customerName?: string | null; 
  hierarchy?: boolean; 
};