// web/src/__generated__/schemas/CustomerActorNewCustomerResponse.ts
/**
 * ⚠️ THIS FILE IS AUTO-GENERATED. DO NOT EDIT MANUALLY.
 * All manual edits will be lost when this file is regenerated.
 */

import type { AuthResponse } from './AuthResponse';

export type CustomerActorNewCustomerResponse = { 
  customerId: string; 
  name: string; 
  realmId: string; 
  realmName: string; 
  token: AuthResponse; 
  activationPending?: boolean; 
  alias?: string | null; 
};